[SPARK-59154][ML] Narrow optimization to decision tree models
