+++ mllib/src/main/scala/org/apache/spark/ml/classification/GBTClassifier.scala
@@ -32,7 +32,6 @@ import org.apache.spark.ml.util.DatasetUtils.extractInstances
 import org.apache.spark.ml.util.DefaultParamsReader.Metadata
 import org.apache.spark.ml.util.Instrumentation.instrumented
 import org.apache.spark.mllib.tree.configuration.{Algo => OldAlgo}
-import org.apache.spark.mllib.tree.loss.{ClassificationLoss => OldClassificationLoss}
 import org.apache.spark.mllib.tree.model.{GradientBoostedTreesModel => OldGBTModel}
 import org.apache.spark.sql._
 import org.apache.spark.sql.functions._
@@ -242,7 +241,6 @@ class GBTClassificationModel private[ml](
     @Since("1.6.0") override val numFeatures: Int,
     @Since("2.2.0") override val numClasses: Int)
   extends ProbabilisticClassificationModel[Vector, GBTClassificationModel]
-  with ProbabilisticClassificationModelTransform[Vector, GBTClassificationModel]
   with GBTClassifierParams with TreeEnsembleModel[DecisionTreeRegressionModel]
   with MLWritable with Serializable {
 
@@ -311,37 +309,6 @@ class GBTClassificationModel private[ml](
         outputSchema($(leafCol)).metadata)
     } else {
       outputData
-    }
-  }
-
-  override protected def predictRawFunction: Vector => Vector = {
-    val localRootNodes = _trees.map(_.rootNode)
-    val localTreeWeights = _treeWeights.clone()
-    features => GBTClassificationModel.predictRaw(features, localRootNodes, localTreeWeights)
-  }
-
-  override protected def raw2probabilityFunction: Vector => Vector = {
-    val localLoss = loss
-    rawPrediction => GBTClassificationModel.raw2probability(rawPrediction, localLoss)
-  }
-
-  override protected def predictionFunction: Vector => Double = {
-    val localRootNodes = _trees.map(_.rootNode)
-    val localTreeWeights = _treeWeights.clone()
-    if (isDefined(thresholds)) {
-      val localThresholds = getThresholds.clone()
-      val localLoss = loss
-      features => {
-        val rawPrediction =
-          GBTClassificationModel.predictRaw(features, localRootNodes, localTreeWeights)
-        val probability = GBTClassificationModel.raw2probability(rawPrediction, localLoss)
-        ProbabilisticClassificationModel.probability2prediction(probability, localThresholds)
-      }
-    } else {
-      features => {
-        val margin = GBTClassificationModel.margin(features, localRootNodes, localTreeWeights)
-        if (margin > 0.0) 1.0 else 0.0
-      }
     }
   }
 
@@ -435,36 +402,6 @@ class GBTClassificationModel private[ml](
 
 @Since("2.0.0")
 object GBTClassificationModel extends MLReadable[GBTClassificationModel] {
-
-  private def margin(
-      features: Vector,
-      rootNodes: Array[Node],
-      treeWeights: Array[Double]): Double = {
-    var prediction = 0.0
-    var i = 0
-    while (i < rootNodes.length) {
-      prediction += rootNodes(i).predictImpl(features).prediction * treeWeights(i)
-      i += 1
-    }
-    prediction
-  }
-
-  private def predictRaw(
-      features: Vector,
-      rootNodes: Array[Node],
-      treeWeights: Array[Double]): Vector = {
-    val prediction = margin(features, rootNodes, treeWeights)
-    Vectors.dense(-prediction, prediction)
-  }
-
-  private def raw2probability(
-      rawPrediction: Vector,
-      loss: OldClassificationLoss): Vector = {
-    val probability = rawPrediction.copy.toDense
-    probability.values(0) = loss.computeProbability(probability.values(0))
-    probability.values(1) = 1.0 - probability.values(0)
-    probability
-  }
 
   private val numFeaturesKey: String = "numFeatures"
   private val numTreesKey: String = "numTrees"
